Add basic Dovecot container.
Add the basics for what will eventually be the Dovecot container. 10-master.conf, 10-ssl.conf, and dovecot-sql.conf.ext have to be modified to allow Dovecot to interact with the database and communicate with the Postfix container. SSL configuration will probably be left for after a working database connection and a communication with Postfix is successfully established.